Move PageBox to its own module
