+++ src/components/common.js
@@ -7,8 +7,6 @@ import Link from 'src/components/common/Link'
 import { icon } from 'src/components/icons'
 import { MiniSortable } from 'src/components/table'
 import scienceBackground from 'src/images/science-background.jpg'
-import { isRadX } from 'src/libs/brand-utils'
-import colors from 'src/libs/colors'
 import { withErrorReporting } from 'src/libs/error'
 import * as Utils from 'src/libs/utils'
 
@@ -24,21 +22,6 @@ export * from 'src/components/common/spinners'
 export * from 'src/components/common/Switch'
 export { Clickable, Link }
 
-
-export const PageBoxVariants = {
-  LIGHT: 'light'
-}
-
-export const PageBox = ({ children, variant, style = {}, ...props }) => {
-  return div(_.merge({
-    style: {
-      margin: '1.5rem', padding: '1.5rem 1.5rem 0', minHeight: 125, flex: 'none', zIndex: 0,
-      ...Utils.switchCase(variant,
-        [PageBoxVariants.LIGHT, () => ({ backgroundColor: colors.light(isRadX() ? 0.3 : 1), margin: 0, padding: '3rem 3rem 1.5rem' })],
-        [Utils.DEFAULT, () => ({})]), ...style
-    }
-  }, props), [children])
-}
 
 export const backgroundLogo = img({
   src: scienceBackground,
