+++ src/pages/FeaturePreviews.js
@@ -1,8 +1,9 @@
 import _ from 'lodash/fp'
 import { Fragment, useCallback, useEffect, useState } from 'react'
 import { div, h, h2, p, span } from 'react-hyperscript-helpers'
-import { Checkbox, Link, PageBox, spinnerOverlay } from 'src/components/common'
+import { Checkbox, Link, spinnerOverlay } from 'src/components/common'
 import FooterWrapper from 'src/components/FooterWrapper'
+import { PageBox } from 'src/components/PageBox'
 import { SimpleFlexTable } from 'src/components/table'
 import TopBar from 'src/components/TopBar'
 import { isFeaturePreviewEnabled, toggleFeaturePreview, useAvailableFeaturePreviews } from 'src/libs/feature-previews'
