+++ src/pages/groups/List.js
@@ -1,13 +1,14 @@
 import _ from 'lodash/fp'
 import { Fragment, useEffect, useState } from 'react'
 import { a, div, h, h2 } from 'react-hyperscript-helpers'
-import { ButtonPrimary, ClipboardButton, DeleteConfirmationModal, HeaderRenderer, IdContainer, Link, PageBox, PageBoxVariants, spinnerOverlay } from 'src/components/common'
+import { ButtonPrimary, ClipboardButton, DeleteConfirmationModal, HeaderRenderer, IdContainer, Link, spinnerOverlay } from 'src/components/common'
 import FooterWrapper from 'src/components/FooterWrapper'
 import { AdminNotifierCheckbox } from 'src/components/group-common'
 import { icon } from 'src/components/icons'
 import { DelayedSearchInput, ValidatedInput } from 'src/components/input'
 import LeaveResourceModal from 'src/components/LeaveResourceModal'
 import Modal from 'src/components/Modal'
+import { PageBox, PageBoxVariants } from 'src/components/PageBox'
 import { ariaSort } from 'src/components/table'
 import TopBar from 'src/components/TopBar'
 import { Ajax } from 'src/libs/ajax'
